knowledge_base/requirements.txt
# Core dependencies
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic==2.11.9
pydantic-settings==2.11.0

# Document processing
PyPDF2==3.0.1
python-multipart==0.0.20
aiofiles==24.1.0
aiohttp==3.11.10
requests==2.32.5
beautifulsoup4==4.14.2
python-magic==0.4.27

# Async and queue management
celery==5.5.3
redis==5.2.1

# Vector operations and ML
numpy==2.3.3
sentence-transformers==5.1.1
chromadb==1.1.0
faiss-cpu==1.12.0

# AWS services (optional)
boto3==1.40.44
opensearch-py==3.0.0

# Database
sqlalchemy==2.0.43
alembic==1.16.5
asyncpg==0.30.0

# Configuration and environment
python-dotenv==1.1.1
PyYAML==6.0.3


# Logging and monitoring
loguru==0.7.3

# Development and testing
pytest==8.4.2
pytest-asyncio==1.2.0
httpx==0.28.1